Remove map related code in geo data scripts
