+++ scripts/localization
@@ -42,26 +42,10 @@ function sync_localizations {
 function update_relay_locations_pot {
     log_header "Retrieving relay locations from server list and translating by using map data"
     pushd ../gui/scripts
-    # Download geo data
-    curl -L -O https://www.naturalearthdata.com/http//www.naturalearthdata.com/download/50m/cultural/ne_50m_admin_0_countries.zip
-    curl -L -O https://www.naturalearthdata.com/http//www.naturalearthdata.com/download/50m/cultural/ne_50m_admin_1_states_provinces_lines.zip
-    curl -L -O https://www.naturalearthdata.com/http//www.naturalearthdata.com/download/10m/cultural/ne_10m_populated_places.zip
-
-    unzip ne_50m_admin_0_countries.zip -d ne_50m_admin_0_countries/
-    unzip ne_50m_admin_1_states_provinces_lines.zip -d ne_50m_admin_1_states_provinces_lines/
-    unzip ne_10m_populated_places.zip -d ne_10m_populated_places/
 
     # Add translations from geo data
-    python3 extract-geo-data.py
-    python3 integrate-into-app.py
-
-    # Remove geo data
-    rm ne_10m_populated_places.zip \
-        ne_50m_admin_0_countries.zip \
-        ne_50m_admin_1_states_provinces_lines.zip
-    rm -r ne_10m_populated_places ne_50m_admin_0_countries ne_50m_admin_1_states_provinces_lines
-
-    git restore ../assets
+    python3 fetch-relay-locations.py
+    python3 integrate-relay-locations.py
 
     popd
 }
